prototypes/paper-grader/requirements.txt
flask==2.3.3
opencv-python==4.8.0.76
pytesseract==0.3.10
numpy==1.25.2
pandas==2.1.0
matplotlib==3.7.2
Pillow==10.0.0
scikit-learn==1.3.0
scikit-image==0.21.0
pyzbar==0.1.9
pdf2image==1.16.3
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
